+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Enhancements:
 
+* Add API to destroy and store documents
+* Add API to manage photos and avatars
 * Enable API web navigation
 * Enhance UI of API's Settings to add comprehension and documentation
 * Improve trim string middleware to not trim password text